refactor: split up test client a bit
