+++ packages/api-postgres/src/testUtils/TestClient.ts
@@ -1,15 +1,10 @@
-import { Application } from 'express';
 import request, { Test } from 'supertest';
-import app from '../config/app';
 import { UpdatableItemFields, Program } from '../common/types';
+import BaseTestClient from './BaseTestClient';
 
-class TestClient {
-    private app: Application;
-    private accessToken: string;
-    private refreshToken: string;
-
+class TestClient extends BaseTestClient {
     constructor() {
-        this.app = app;
+        super();
     }
 
     public getIndexRoute(): Test {
@@ -76,18 +71,6 @@ class TestClient {
 
     public logoutUser(): void {
         this.clearTokens();
-    }
-
-    public async doLogin(email: string, password: string): Promise<void> {
-        this.clearTokens();
-        const response = await request(this.app)
-            .post('/api/auth/login')
-            .set('Content-Type', 'application/json')
-            .send({ email, password });
-
-        this.accessToken = response.body.payload.accessToken;
-        const rToken = this.getRefreshTokenFromHeaders(response.header['set-cookie']);
-        this.refreshToken = rToken;
     }
 
     public createItem(itemData: any): Test {
@@ -169,26 +152,6 @@ class TestClient {
             .get(`/api/clientrequests/${id}`)
             .set('Cookie', [`qid=${this.refreshToken}`])
             .set('Authorization', `Bearer ${this.accessToken}`);
-    }
-
-    private clearTokens(): void {
-        this.accessToken = '';
-        this.refreshToken = '';
-    }
-
-    private getRefreshTokenFromHeaders(headers: string[]): string {
-        let refreshToken = '';
-
-        if (!headers || headers.length === 0) {
-            return refreshToken;
-        }
-
-        const cookies = headers[0].split(';');
-        const qid = cookies.find((cookie: string) => cookie.indexOf('qid=') === 0);
-        if (qid) {
-            refreshToken = qid.split('=')[1];
-        }
-        return refreshToken;
     }
 }
